requirements.txt
fastapi[all]
jose
passlib
bcrypt
spacy
sentence-transformers
pydantic
google-generativeai
sqlalchemy
faiss-cpu
pypdf2
beautifulsoup4
lxml
google-search-results
